Rename src -> autocorpus so the package name matches the tool
