+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
       - id: check-merge-conflict
       - id: debug-statements
       - id: trailing-whitespace
-        exclude: ^src/IAO_dicts/
+        exclude: ^autocorpus/IAO_dicts/
       - id: end-of-file-fixer
       - id: pretty-format-json
         args: [--autofix, --indent, '4', --no-sort]
